feat(demo): add new demo scenes

+++ Assets/cARdefender/Assets/Enemies/Generic Enemies/Scripts/IHittableEnemy.cs
@@ -16,6 +16,6 @@ public interface IHittableEnemy
     /// </summary>
     /// <param name="damage">How much damage to deal.</param>
     
-    void OnTakeDamage(int damage);
+    void OnTakeDamage(float damage);
 
 }
